potentially relevant if soil parameters are unrealistic
Catch floating point exception when calculated unsaturated hydraulic conductivity is zero (used for calculation of percolation) as a result of unrealistic soil parameters. In case floating point exceptions are not caught by the compiler (which is usually the case), this would lead to some (potentially undetected) NaN values in the simulation of soil moisture.