ENH: comput. irrelevant: variable 'dtot'
Variable dtot should be the 'total number of days simulated so far' but actually it was the 'total number of days that will have been simulated at the end of the current year so far'. Now dtot will be incremented after every simulation day and represents the current total simulation day number.